Moved structs and enums to separate files
