--- QuizApp/QuizApp/Models/Quiz.swift
@@ -1,60 +1,0 @@
-import UIKit
-
-enum Difficulty: String {
-
-    case easy = "EASY"
-    case normal = "NORMAL"
-    case hard = "HARD"
-
-}
-
-enum CategoryType: CaseIterable {
-
-    case sport
-    case movies
-    case music
-    case geography
-
-}
-
-struct Category: Equatable {
-
-    var name: String {
-        switch type {
-        case .sport:
-            return "Sport"
-        case .movies:
-            return "Movies"
-        case .music:
-            return "Music"
-        case .geography:
-            return "Geography"
-        }
-    }
-
-    let type: CategoryType
-    let color: UIColor
-
-}
-
-extension Category {
-
-    init(from type: CategoryType) {
-        self.type = type
-        color = .white
-    }
-
-    init(from model: CategoryModel) {
-        switch model {
-        case .sport:
-            self.init(type: CategoryType.sport, color: .sportColor)
-        case .movies:
-            self.init(type: CategoryType.movies, color: .moviesColor)
-        case .music:
-            self.init(type: CategoryType.music, color: .musicColor)
-        case .geography:
-            self.init(type: CategoryType.geography, color: .geographyColor)
-        }
-    }
-
-}
