+++ QuizApp/QuizApp/Network/NetworkClients/QuizNetworkClient.swift
@@ -19,32 +19,3 @@ class QuizNetworkClient: QuizNetworkClientProtocol {
     }
 
 }
-
-struct QuizNetworkDataModel: Decodable {
-
-    let id: Int
-    let name: String
-    let description: String
-    let category: CategoryNetworkDataModel
-    let difficulty: DifficultyNetworkDataModel
-    let imageUrl: String
-    let numberOfQuestions: Int
-
-}
-
-enum CategoryNetworkDataModel: String, Decodable {
-
-    case sport = "SPORT"
-    case movies = "MOVIES"
-    case music = "MUSIC"
-    case geography = "GEOGRAPHY"
-
-}
-
-enum DifficultyNetworkDataModel: String, Decodable {
-
-    case easy = "EASY"
-    case normal = "NORMAL"
-    case hard = "HARD"
-
-}
