+++ QuizApp/QuizApp/Network/NetworkClients/UserNetworkClient.swift
@@ -29,17 +29,3 @@ class UserNetworkClient: UserNetworkClientProtocol {
     }
 
 }
-
-struct UserInfoNetworkDataModel: Decodable {
-
-    let id: Int
-    let email: String
-    let name: String
-
-}
-
-struct UserInfoNetworkRequestModel: Encodable {
-
-    let name: String
-
-}
