+++ QuizApp/QuizApp/Network/NetworkDataSources/QuizNetworkDataSource.swift
@@ -20,46 +20,3 @@ class QuizNetworkDataSource: QuizNetworkDataSourceProtocol {
     }
 
 }
-
-struct QuizResponseDataModel {
-
-    let id: Int
-    let name: String
-    let description: String
-    let category: CategoryDataModel
-    let difficulty: DifficultyDataModel
-    let imageUrl: String
-    let numberOfQuestions: Int
-
-}
-
-extension QuizResponseDataModel {
-
-    init(_ quiz: QuizNetworkDataModel) {
-        id = quiz.id
-        name = quiz.name
-        description = quiz.description
-        category = CategoryDataModel(rawValue: quiz.category.rawValue)!
-        difficulty = DifficultyDataModel(rawValue: quiz.difficulty.rawValue)!
-        imageUrl = quiz.imageUrl
-        numberOfQuestions = quiz.numberOfQuestions
-    }
-
-}
-
-enum CategoryDataModel: String {
-
-    case sport = "SPORT"
-    case movies = "MOVIES"
-    case music = "MUSIC"
-    case geography = "GEOGRAPHY"
-
-}
-
-enum DifficultyDataModel: String {
-
-    case easy = "EASY"
-    case normal = "NORMAL"
-    case hard = "HARD"
-
-}
