+++ QuizApp/QuizApp/Repositories/QuizRepository.swift
@@ -20,46 +20,3 @@ class QuizRepository: QuizRepositoryProtocol {
     }
 
 }
-
-struct QuizRepoModel {
-
-    let id: Int
-    let name: String
-    let description: String
-    let category: CategoryRepoModel
-    let difficulty: DifficultyRepoModel
-    let imageUrl: String
-    let numberOfQuestions: Int
-
-}
-
-extension QuizRepoModel {
-
-    init(_ quiz: QuizResponseDataModel) {
-        id = quiz.id
-        name = quiz.name
-        description = quiz.description
-        category = CategoryRepoModel(rawValue: quiz.category.rawValue)!
-        difficulty = DifficultyRepoModel(rawValue: quiz.difficulty.rawValue)!
-        imageUrl = quiz.imageUrl
-        numberOfQuestions = quiz.numberOfQuestions
-    }
-
-}
-
-enum CategoryRepoModel: String {
-
-    case sport = "SPORT"
-    case movies = "MOVIES"
-    case music = "MUSIC"
-    case geography = "GEOGRAPHY"
-
-}
-
-enum DifficultyRepoModel: String {
-
-    case easy = "EASY"
-    case normal = "NORMAL"
-    case hard = "HARD"
-
-}
