+++ QuizApp/QuizApp/Repositories/UserRepository.swift
@@ -48,33 +48,3 @@ class UserRepository: UserRepositoryProtocol {
     }
 
 }
-
-struct LoginResponseRepoModel {
-
-    var accessToken: String
-
-}
-
-extension LoginResponseRepoModel {
-
-    init(_ responseDataModel: LoginResponseDataModel) {
-        accessToken = responseDataModel.accessToken
-    }
-
-}
-
-struct UserInfoRepoModel {
-
-    let username: String
-    let name: String
-
-}
-
-extension UserInfoRepoModel {
-
-    init(_ userInfo: UserInfoDataModel) {
-        username = userInfo.email
-        name = userInfo.name
-    }
-
-}
