+++ QuizApp/QuizApp/UseCase/LoginUseCase.swift
@@ -17,17 +17,3 @@ class LoginUseCase: LoginUseCaseProtocol {
     }
 
 }
-
-struct LoginResponseModel {
-
-    let accessToken: String
-
-}
-
-extension LoginResponseModel {
-
-    init(_ loginResponse: LoginResponseRepoModel) {
-        accessToken = loginResponse.accessToken
-    }
-
-}
